refactor: 공통 UI 컴포넌트 위치 수정
