+++ app/build.gradle.kts
@@ -20,7 +20,7 @@ android {
         applicationId = "desktop.hambug"
         minSdk = 30
         targetSdk = 36
-        versionCode = 1
+        versionCode = 100
         versionName = "1.0"
 
         testInstrumentationRunner = "androidx.test.runner.AndroidJUnitRunner"
