+++ app/src/main/java/desktop/hambug/presentation/community/BoardWriteScreen.kt
@@ -49,10 +49,10 @@ import androidx.navigation.NavHostController
 import coil3.compose.AsyncImage
 import desktop.hambug.domain.model.Category
 import desktop.hambug.presentation.community.component.RequiredFieldTitle
-import desktop.hambug.presentation.ui.component.CustomContentTextField
-import desktop.hambug.presentation.ui.component.CustomSnackbar
-import desktop.hambug.presentation.ui.component.CustomTitleTextField
-import desktop.hambug.presentation.ui.component.HambugLoadingIndicator
+import desktop.hambug.presentation.component.CustomContentTextField
+import desktop.hambug.presentation.component.CustomSnackbar
+import desktop.hambug.presentation.component.CustomTitleTextField
+import desktop.hambug.presentation.component.HambugLoadingIndicator
 import desktop.hambug.presentation.ui.icon.AppIcons
 import desktop.hambug.presentation.ui.icon.appicons.BackDetail
 import desktop.hambug.presentation.ui.icon.appicons.Camera
