+++ app/src/main/java/desktop/hambug/presentation/component/CustomSnackbar.kt
@@ -1,4 +1,4 @@
-package desktop.hambug.presentation.ui.component
+package desktop.hambug.presentation.component
 
 import androidx.compose.foundation.Image
 import androidx.compose.foundation.background
