+++ app/src/main/java/desktop/hambug/presentation/component/ForceUpdateDialog.kt
@@ -1,4 +1,4 @@
-package desktop.hambug.presentation.ui.component
+package desktop.hambug.presentation.component
 
 import androidx.compose.foundation.background
 import androidx.compose.foundation.clickable
