+++ app/src/main/java/desktop/hambug/presentation/my/MyActivityScreen.kt
@@ -47,7 +47,7 @@ import coil3.compose.AsyncImage
 import desktop.hambug.domain.model.MyBoard
 import desktop.hambug.domain.model.MyComment
 import desktop.hambug.presentation.component.EmptyStateView
-import desktop.hambug.presentation.ui.component.HambugLoadingIndicator
+import desktop.hambug.presentation.component.HambugLoadingIndicator
 import desktop.hambug.presentation.ui.icon.AppIcons
 import desktop.hambug.presentation.ui.icon.appicons.BackDetail
 import desktop.hambug.presentation.ui.icon.appicons.Comment
