+++ app/src/main/java/desktop/hambug/presentation/my/component/NicknameUpdateDialog.kt
@@ -25,7 +25,7 @@ import androidx.compose.ui.unit.dp
 import androidx.compose.ui.window.Dialog
 import desktop.hambug.R
 import desktop.hambug.presentation.my.NicknameValidationState
-import desktop.hambug.presentation.ui.component.HambugLoadingIndicator
+import desktop.hambug.presentation.component.HambugLoadingIndicator
 import desktop.hambug.presentation.ui.theme.HambugTheme
 
 @Composable
