+++ app/src/main/java/desktop/hambug/presentation/my/component/UserRemoveDialog.kt
@@ -23,7 +23,7 @@ import androidx.compose.ui.unit.dp
 import androidx.compose.ui.window.Dialog
 import androidx.compose.ui.window.DialogProperties
 import desktop.hambug.R
-import desktop.hambug.presentation.ui.component.HambugLoadingIndicator
+import desktop.hambug.presentation.component.HambugLoadingIndicator
 import desktop.hambug.presentation.ui.theme.HambugTheme
 
 @Composable
